asset centrality and relative value combined.
This blend of asset centrality and relative value enables the identification of whether there exists a market bubble or not within the asset group. Additionally, it helps discern whether the bubble, if present, is undergoing a rising or falling trend based on the relative value. Consequently, three distinct portfolios of assets are constructed: one for scenarios where there is no bubble, another for situations where the bubble is on the rise, and a third for instances where the bubble is on the decline.